Add files via upload
This is the initially trained model to ensure that the model is coded correctly. It is iterated a minimum of 50x per version (with additional repeat runs when needed). The losses are printed out, to ensure that the model isn't being over-fitted and to determine if additional iterations of training need to be performed on the model. It also provides an indication of how the modelling is running and whether training is done.